pyproject.toml
[tool.poetry]
name = "Python AI Hub"
version = "0.1.0"
license = "MIT"
description = ""
authors = ["Yevhen Kuzminov <kyzminov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
pillow = "^10.3.0"
environs = {extras = ["django"], version = "^11.0.0"}
factory-boy = "^3.3.0"
faker = "^24.7.1"

[tool.poetry.dev-dependencies]

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
